Move RequestPasswordResetRequest to email folder
