+++ api/src/test/java/com/ford/labs/retroquest/api/TeamApiTest.java
@@ -19,6 +19,7 @@ package com.ford.labs.retroquest.api;
 
 import com.ford.labs.retroquest.api.setup.ApiTestBase;
 import com.ford.labs.retroquest.column.ColumnRepository;
+import com.ford.labs.retroquest.email.RequestPasswordResetRequest;
 import com.ford.labs.retroquest.team.*;
 import com.ford.labs.retroquest.team.password.PasswordResetToken;
 import com.ford.labs.retroquest.team.password.PasswordResetTokenRepository;
